requirements.txt
langchain
python-dotenv
openai
qdrant-client
tiktoken
streamlit
